ai-agent-rag-solution/requirements.txt
# Core dependencies
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# LLM and AI
litellm==1.48.7
openai==1.54.3

# RAG and Vector Store
chromadb==0.4.22
sentence-transformers==2.2.2
langchain==0.1.0
langchain-community==0.0.10

# PDF Processing
pypdf2==3.0.1
pymupdf==1.23.8
pdfplumber==0.10.3

# Embeddings
numpy==1.24.3
torch==2.1.2
transformers==4.36.2

# Session Management
redis==5.0.1
uuid==1.30

# Utilities
tqdm==4.66.1
requests==2.31.0
